recursive imports working all the way down now, most sections sketched out - radiative transfer will need some additional work while building into code base
