+++ isofit/configs/__init__.py
@@ -1,2 +1,2 @@
-from .configs import BaseConfigSection
-from .configs import Config
+#from .configs import BaseConfigSection
+#from .configs import Config
