+++ isofit/configs/sections/__init__.py
@@ -1,4 +1,7 @@
-from .input_config import InputConfig
-from .output_config import OutputConfig
-from .forward_model_config import ForwardModelConfig
-from .instrument_config import InstrumentConfig
+#from .input_config import InputConfig
+#from .output_config import OutputConfig
+#from .forward_model_config import ForwardModelConfig
+#from .instrument_config import InstrumentConfig
+#from .radiative_transfer_config import RadiativeTransferModelConfig
+#from .radiative_transfer_config import RadiativeTransferConfig
+#from .radiative_transfer_config import RadiativeTransferUnknownsConfig
