+++ isofit/configs/sections/input_config.py
@@ -2,7 +2,7 @@
 
 
 from typing import Dict, List, Type
-from isofit.configs import BaseConfigSection
+from isofit.configs.base_config import BaseConfigSection
 import os
 
 class InputConfig(BaseConfigSection):
